requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-slugify>=8.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
python-dotenv>=1.0.0